Add CLAUDE.md with project overview and development guidelines
