+++ src/app/sync/sync.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SyncService } from './sync.service';
+
+describe('SyncService', () => {
+  let service: SyncService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SyncService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
